+++ scripts/project_meta.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+
+PROJECT_NAME="crona"
+PROJECT_REPO="webxsid/crona"
+PROJECT_BIN_PREFIX="crona"